KA-269 Rename utils to .ts and update tests
